LearningSchedules

Documentation for LearningSchedules.

LearningSchedules.Burnin

LearningSchedules.BurninHyperbolic

LearningSchedules.Linear

LearningSchedules.next_rate

LearningSchedules.next_rate!

LearningSchedules.Burnin — Type

Burnin{T} <: TimeBased{T}
Burnin(min::T, max::T, inflate::T, decay::T)

A learning rate schedule with exponential inflation and exponential decay stages. The rate starts at min, inflates exponentially to max, then decays exponentially to min.

Arguments

min::T: The minimum learning rate.

max::T: The maximum learning rate.

inflate::T: The inflation factor during stage 1.

decay::T: The decay factor during stage 2 (starts after max is reached).

LearningSchedules.BurninHyperbolic — Type

BurninHyperbolic{T} <: TimeBased{T}
BurninHyperbolic(min::T, max::T, inflate::T, decay::T, floor::T)

A learning rate schedule with exponential inflation and hyperbolic decay stages. The rate starts at min, inflates exponentially to max, then decays hyperbolically to min.

Arguments

min::T: The minimum learning rate.

max::T: The maximum learning rate.

inflate::T: The inflation factor during stage 1.

decay::T: The decay factor during stage 2 (starts after max is reached).

floor::T: idk ask Ben or look at the code lol

LearningSchedules.Linear — Type

Linear{T} <: StepBased{T}
Linear(initial::T, final::T, steps::Int)

A linear learning rate schedule, decaying from initial to final over steps iterations.

Arguments

initial::T: The initial learning rate.

final::T: The final learning rate.

steps::Int: The number of steps to decay over.

LearningSchedules.next_rate! — Method

next_rate!(schedule::Stateful{<:LearningRateSchedule})

Update the schedule and return the next rate.

See also: next_rate

Examples

julia> linear_with_state = Linear(1.0, 0.0, 10)(); # stateful

julia> next_rate!(linear_with_state)
1.0

julia> next_rate!(linear_with_state)
0.9

LearningSchedules.next_rate — Method

next_rate(schedule::Stateful{<:LearningRateSchedule})

See also: next_rate!

Examples

julia> linear_with_state = Linear(1.0, 0.0, 10)(); # stateful

julia> next_rate(linear_with_state)
1.0

julia> next_rate!(linear_with_state)
1.0

julia> next_rate(linear_with_state)
0.9